requirements.txt
lxml==5.3.0
numpy==2.0.1
pandas==2.2.2
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
